requirements.txt
flask
gunicorn
requests
playwright.sync_api